Fix drag while drawing not working.
Because handle_draw_mode() was overwriting a state even when was
not handled nothing.

Fix: the drag_start_src state are only written by handle_draw_mode
if it is handling a event.

This need to be checked in every future child View2D handle. A
View2D handle state cannot be writen by handle that handle nothing.